view sales-crated, sales tests running

+++ database/migrations/2019_10_18_213231_create_products_table.php
@@ -17,9 +17,9 @@ class CreateProductsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('barcode')->nullable();
-            $table->float('price', 3, 2);
+            $table->decimal('price', 8, 2);
             $table->timestamps();
-            
+
             $table->unsignedBigInteger('category_id');
         });
     }
